requirements.txt
requests
pandas
numpy
beautifulsoup4
spacy
lxml
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.3
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
executing==2.2.1
future==0.18.2
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
jiter==0.11.0
jupyter_client==8.6.3
jupyter_core==5.8.1
macholib==1.15.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
openai==2.1.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
six==1.15.0
sniffio==1.3.1
stack-data==0.6.3
tiktoken==0.11.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
zipp==3.23.0

